+++ src/autopas/containers/verletListsCellBased/verletListsCells/neighborLists/VLCAllCellsNeighborList.h
@@ -34,7 +34,7 @@ class VLCAllCellsNeighborList : public VLCNeighborListInterface<Particle_T> {
   /**
    * Helper type definition. Pair of particle and neighbor list for SoA layout.
    */
-  using SoAPairOfParticleAndList = std::pair<size_t, std::vector<size_t, autopas::AlignedAllocator<size_t>>>;
+  using SoAPairOfParticleAndList = std::pair<uint32_t, std::vector<uint32_t, autopas::AlignedAllocator<uint32_t>>>;
 
   /**
    * Helper type definition. Vector of cells, for each cell a vector of neighbors.
@@ -309,7 +309,7 @@ class VLCAllCellsNeighborList : public VLCNeighborListInterface<Particle_T> {
         size_t currentParticleGlobalIndex = particlePtrToIndex.at(particlePtr);
 
         // create SoA neighbor list for current particle
-        std::vector<size_t, autopas::AlignedAllocator<size_t>> currentSoANeighborList{};
+        std::vector<uint32_t, autopas::AlignedAllocator<uint32_t>> currentSoANeighborList{};
         currentSoANeighborList.reserve(neighbors.size());
 
         // fill the SoA neighbor list with the indices of the particles from the corresponding AoS neighbor list
